.github/scripts/check_project_fos_precision/requirements.txt
asn1==2.7.0
async-generator==1.10
attrs==21.4.0
beautifulsoup4==4.11.1
blinker==1.6.3
certifi==2024.2.2
cffi==1.15.0
chardet==5.1.0
click==6.7
configobj==5.0.8
cryptography==37.0.2
Deprecated==1.2.13
enum-compat==0.0.3
Flask==1.0.4
Flask-WTF==0.14.3
gitdb==4.0.11
GitPython==3.1.43
gunicorn==20.1.0
h11==0.13.0
icalendar==5.0.12
idna==3.7
iniconfig==1.1.1
itsdangerous==0.24
Jinja2==2.11.3
ldap3==2.9.1
MarkupSafe==2.0.1
numpy==1.26.4
outcome==1.1.0
packaging==21.3
pandas==2.2.2
pluggy==1.0.0
prometheus-client==0.20.0
py==1.11.0
pyasn1==0.5.1
pyasn1-modules==0.2.8
pycparser==2.21
PyGithub==1.57
PyJWT==2.6.0
PyNaCl==1.5.0
pyOpenSSL==22.0.0
pyparsing==3.0.7
PySocks==1.7.1
pytest==7.1.1
pytest-mock==3.7.0
python-dateutil==2.8.2
python-gnupg==0.5.2
python-ldap==3.3.1
pytz==2024.1
PyYAML==6.0.1
requests==2.25.1
selenium==4.1.3
six==1.16.0
smmap==5.0.1
sniffio==1.2.0
sortedcontainers==2.4.0
soupsieve==2.3.2.post1
tomli==2.0.1
tqdm==4.64.0
trio==0.20.0
trio-websocket==0.9.2
tzdata==2024.1
urllib3==1.26.6
webdriverdownloader==1.1.0.3
Werkzeug==0.15.6
wrapt==1.14.1
wsproto==1.1.0
WTForms==3.0.1
xmltodict==0.13.0
